review: rename dns_cache to dnscache
