+++ dnscache/storage.go
@@ -1,4 +1,4 @@
-package dns_cache
+package dnscache
 
 import (
 	"net"
